Fix SDL controller disconnection.
https://wiki.libsdl.org/SDL_ControllerDeviceEvent

Confusingly, the `which` variable does not mean the same thing for
`SDL_CONTROLLERDEVICEADDED` as it does for
`SDL_CONTROLLERDEVICEREMOVED`: in the former, it's the device index,
which in the latter, it's the instance ID.

Because of this, controllers weren't being removed when they should
have been because their IDs were never correct.

To remedy this, the instance ID is now properly used for both
connections and disconnections.