Fix gamePadMappings memory leak
This array was allocated using `new` even if the mod loader was disabled,
and the size used for the allocation was never checked to be valid.

Only allocate/free when mod loader is enabled, and add a size check.

ASan on Linux showed a 1 byte leak when `gamePadCount==0`, who knows
what was happening on other platforms.